appveyor.yml: use Miniconda 3.5 to build our Python 3.6 packages
This is how the conda-smithy system *used* to work, but at some point Appveyor
deployed Miniconda 3.6 and so presumably conda-smithy updated to reflect its
availability. *However*, astropy needs to use an older image because VS2015
Update 3 includes a compiler bug that breaks its build, and this older image
doesn't have Miniconda 3.6. So, we need to keep on using Miniconda 3.5 here.